# docs/swagger.yaml
basePath: /
definitions:
  queue.Message:
    properties:
      value:
        type: string
    type: object
  response.Error:
    properties:
      error:
        type: string
      info:
        type: string
    type: object
  response.OK:
    properties:
      ok:
        type: boolean
    type: object
  response.Ping:
    properties:
      ping:
        type: string
    type: object
  response.QueueCreateResponse:
    properties:
      info:
        type: string
      name:
        type: string
      ok:
        type: boolean
      presize:
        type: integer
    type: object
  response.QueueDeleteResponse:
    properties:
      info:
        type: string
      name:
        type: string
      ok:
        type: boolean
    type: object
  response.QueueInfo:
    properties:
      count:
        type: integer
      head:
        type: integer
      name:
        type: string
      presize:
        type: integer
      size:
        type: integer
      tail:
        type: integer
    type: object
  response.QueueList:
    properties:
      queue_names:
        items:
          type: string
        type: array
    type: object
  response.RootResponse:
    properties:
      author:
        type: string
      docs:
        type: string
      license:
        type: string
      limero:
        type: string
      version:
        type: string
    type: object
host: localhost:7920
info:
  contact: {}
  description: This is a message broker
  license:
    name: MIT license
  title: Limero
  version: "0.1"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Returns information about the limero
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.RootResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/response.Error'
      summary: Limero information
      tags:
      - root
  /msg:
    get:
      consumes:
      - application/json
      description: Get message from the queue by name
      parameters:
      - description: Queue name
        in: query
        name: qname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/queue.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      summary: Get message from the queue
      tags:
      - msg
    post:
      consumes:
      - application/json
      description: Sends a message to queue by name
      parameters:
      - description: Queue name
        in: query
        name: qname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.OK'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      summary: Sends a message to queue
      tags:
      - msg
  /ping:
    get:
      consumes:
      - application/json
      description: Returns information about the limero
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Ping'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/response.Error'
      summary: Limero information
      tags:
      - root
  /queue:
    delete:
      consumes:
      - application/json
      description: Deletes a queue by name
      parameters:
      - description: Queue name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.QueueDeleteResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      summary: Deletes a queue
      tags:
      - queue
    get:
      consumes:
      - application/json
      description: Returns a list of queue names
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.QueueList'
      summary: Queue list
      tags:
      - queue
    put:
      consumes:
      - application/json
      description: Creates a new queue with a name and size
      parameters:
      - description: Queue name
        in: query
        name: name
        required: true
        type: string
      - description: Queue size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.QueueCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Error'
      summary: Creates a new queue
      tags:
      - queue
  /queue/{qname}:
    get:
      consumes:
      - application/json
      description: Returns information about the queue by its name
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.QueueInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Error'
      summary: Information about queue
      tags:
      - queue
swagger: "2.0"